Add Increment: Level-3: Mark as Done
This version of Duke has the ability to mark tasks as done. It also
alerts the user that the list is empty and prompts him or her to enter
again in the event that the user types "list" at the start of the
program (when there is no task to do).